+++ src/structlog/tracebacks.py
@@ -81,10 +81,14 @@ class SyntaxError_:  # noqa: N801
 class Stack:
     """
     Represents an exception and a list of stack frames.
+
+    .. versionchanged:: 25.2.0
+       Added the *exc_notes* field.
     """
 
     exc_type: str
     exc_value: str
+    exc_notes: list[str] = field(default_factory=list)
     syntax_error: SyntaxError_ | None = None
     is_cause: bool = False
     frames: list[Frame] = field(default_factory=list)
@@ -230,6 +234,9 @@ def extract(
         stack = Stack(
             exc_type=safe_str(exc_type.__name__),
             exc_value=safe_str(exc_value),
+            exc_notes=[
+                safe_str(note) for note in getattr(exc_value, "__notes__", ())
+            ],
             is_cause=is_cause,
         )